pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plexus"
version = "0.1.0"
description = "The Universal Signal Nervous System"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["plexus"]
